--- app/(route)/(header)/layout.tsx
@@ -1,13 +1,0 @@
-import React, { ReactNode } from "react";
-import MobileHeader from "@/components/header/MobileHeader";
-
-const TopAppBarLayout = ({ children }: { children: ReactNode }) => {
-  return (
-    <>
-      <MobileHeader />
-      {children}
-    </>
-  );
-};
-
-export default TopAppBarLayout;
