--- app/(route)/(header)/my-artist-event/page.tsx
@@ -1,11 +1,0 @@
-import MyArtistEvent from "./_components/MyArtistEvent";
-
-const MyArtistEventPage = () => {
-  return (
-    <div className="flex flex-col px-20 pb-88 pt-16">
-      <MyArtistEvent />
-    </div>
-  );
-};
-
-export default MyArtistEventPage;
