+++ app/(route)/event/[eventId]/_components/Banner.tsx
