+++ app/(route)/event/[eventId]/_components/KakaoMap.tsx
