+++ app/(route)/event/[eventId]/_components/tab/DescriptionTab.tsx
