+++ app/(route)/event/[eventId]/approve/[editId]/_components/EditBox.tsx
