+++ app/(route)/event/[eventId]/approve/_components/RenderException.tsx
