+++ app/_components/header/MobileHeader.tsx
@@ -33,7 +33,7 @@ const MobileHeader = ({ handleClick }: Props) => {
   return (
     <>
       <header
-        className={`sticky left-0 top-0 z-nav flex h-72 w-full justify-between border-b border-gray-50 bg-white-white px-20 pb-12 pt-36 ${pathname === `/event/${eventId}` ? "pc:hidden" : ""}`}
+        className={`sticky left-0 top-0 z-nav flex h-72 w-full justify-between border-b border-gray-50 bg-white-white px-20 pb-12 pt-36 ${pathname === `/event/${eventId}` || pathname === "/my-artist-event" ? "pc:hidden" : ""}`}
       >
         <button onClick={handleClick || (() => router.back())} className="z-nav">
           <ArrowLeft />
